% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Run Statistical Analysis of Monthly Background Checks of Gun Purchase}
\usage{
analysis(debug = FALSE)
}
\arguments{
\item{debug}{Optional boolean switch to indicate whether interim data is displayed; default is \sQuote{FALSE}}
}
\value{
A \code{data.frame} is returned, contained all different prepared columns.
}
\description{
Run Statistical Analysis of Monthly Background Checks of Gun Purchase
}
\examples{
gs <- analysis()
plot_gunsales(gs)
ggplot_gunsales(gs)
}
\author{
Gregor Aisch and Josh Keller wrote the R code; Dirk Eddelbuettel created and maintains the package.
}
\seealso{
The NY Times article presenting this analsysi undertaken by this package is
at \url{http://www.nytimes.com/interactive/2015/12/10/us/gun-sales-terrorism-obama-restrictions.html?}
}

