% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{gganalysis}
\alias{gganalysis}
\title{Run Statistical Analysis of Monthly Background Checks of Gun Purchase}
\usage{
gganalysis(savePlots = FALSE, saveData = FALSE)
}
\arguments{
\item{savePlots}{Optional boolean switch to indicate whether plots are saved, default is \sQuote{FALSE}}

\item{saveData}{Optional boolean switch to indicate whether csv files are saved, default is \sQuote{FALSE}}
}
\value{
\code{NULL} is returned invisibly
}
\description{
Run Statistical Analysis of Monthly Background Checks of Gun Purchase
}
\examples{
gganalysis()
}
\author{
Gregor Aisch and Josh Keller wrote the R code; Dirk Eddelbuettel created and maintains the package.
}
\seealso{
The NY Times article presenting this analsysi undertaken by this package is
at \url{http://www.nytimes.com/interactive/2015/12/10/us/gun-sales-terrorism-obama-restrictions.html?}
}

