% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot.R
\name{ggplot_gunsales}
\alias{ggplot_gunsales}
\title{ggplot2 plots for gunsales analysis}
\usage{
ggplot_gunsales(df, savePlots = FALSE, savePNG = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} as prepared by the
\code{\link{analysis}} functions.}

\item{savePlots}{A boolean toggle to indicate if the plots are to
be saved in the \code{out/} directory as a single pdf file,
with a default of \code{FALSE}.}

\item{savePNG}{A boolean toggle to indicate if the plots are to
be saved in the \code{out/} directory as individual png files,
with a default of \code{FALSE}.}
}
\value{
\code{NULL} is returned invisibly.
}
\description{
This function creates (ggplot2) plots for gunsales analysis.
}
\details{
In interactive mode, plot display is paused and the user has to
advance by pressing the Return key.
}
\examples{
\dontrun{
  gs <- analysis()
  ggplot_gunsales(gs)
}
}
\author{
Gregor Aisch and Josh Keller wrote the R code; Dirk
    Eddelbuettel created and maintains the package.
}
\seealso{
The NY Times article presenting this analsysi undertaken
    by this package is at
    \url{http://www.nytimes.com/interactive/2015/12/10/us/gun-sales-terrorism-obama-restrictions.html?}
}

