% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_gunsales}
\alias{plot_gunsales}
\title{Base Plots for gunsales analysis}
\usage{
plot_gunsales(df, savePlots = FALSE)
}
\arguments{
\item{df}{A \code{data.frame} as prepared by the
\code{\link{analysis}} functions.}

\item{savePlots}{A boolean toggle to indicate if the plots are to
be saved in the \code{out/} directory, with a default of
\code{FALSE}.}
}
\value{
\code{NULL} is returned invisibly.
}
\description{
This function creates (base graphics) plots for gunsales analysis.
}
\details{
In interactive mode, plot display is paused and the user has to
advance by pressing the Return key.
}
\examples{
\dontrun{
  gs <- analysis()
  plot_gunsales(gs)
}
}
\author{
Gregor Aisch and Josh Keller wrote the R code; Dirk
    Eddelbuettel created and maintains the package.
}
\seealso{
The NY Times article presenting this analsysi undertaken
    by this package is at
    \url{http://www.nytimes.com/interactive/2015/12/10/us/gun-sales-terrorism-obama-restrictions.html?}
}

