% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poptotal}
\alias{poptotal}
\title{US Population Growth data}
\format{A \sQuote{tbl_df} and data frame object with 217 observation of 3 variables}
\source{
US Census Bureau
}
\description{
Montly observations about population growth
}

